extract MembersModal to its own component so that Add Members, Create DM/GM, and ManageMembers components will not have to manage the search and selection of profiles
